third_party/blink/web_tests/platform/mac/external/wpt/mathml/presentation-markup/operators/operator-dictionary-combining-expected.txt

This is a testharness.js-based test.
PASS Spacing around ≠
PASS Spacing around =⃒
FAIL Spacing around |̸ assert_approx_equals: expected 66.65625 +/- 1 but got 26.671875
PASS Spacing around |⃒
Harness: the test ran to completion.
